Settings plugin to share repositories/mirrors and build-logic includes
