+++ build-logic-base/gradle.properties
