+++ build-logic-base/settings.gradle.kts
@@ -1,0 +1,7 @@
+include("settings-plugin")
+
+dependencyResolutionManagement {
+    repositories {
+        gradlePluginPortal()
+    }
+}
