+++ build-logic-commons/gradle.properties
@@ -1,1 +1,15 @@
+org.gradle.jvmargs=-Xmx2500m -XX:MaxMetaspaceSize=512m -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8
+org.gradle.parallel=true
+kotlin.parallel.tasks.in.project=true
+org.gradle.caching=true
+systemProp.org.gradle.kotlin.dsl.caching.buildcache=true
+systemProp.gradle.publish.skip.namespace.check=true
+systemProp.org.gradle.internal.ide.scan=true
+org.gradle.vfs.watch=true
+org.gradle.vfs.verbose=true
+org.gradle.vfs.debug=true
+kotlin.stdlib.default.dependency=false
+gradle.internal.testdistribution.writeTraceFile=true
+systemProp.gradle.internal.testdistribution.writeTraceFile=true
+systemProp.gradle.internal.testdistribution.queryResponseTimeout=PT20S
 systemProp.org.gradle.kotlin.dsl.precompiled.accessors.strict=true
