+++ build-logic/binary-compatibility/src/main/groovy/gradlebuild.binary-compatibility.gradle
@@ -29,24 +29,6 @@ plugins {
     id("gradlebuild.module-identity")
 }
 
-repositories {
-    ['distributions', 'distributions-snapshots'].each { distUrl ->
-        ivy {
-            name 'Gradle distributions'
-            url 'https://services.gradle.org'
-            patternLayout {
-                artifact "/${distUrl}/[module]-[revision]-bin(.[ext])"
-            }
-            metadataSources {
-                artifact()
-            }
-            content {
-                includeModule('gradle', 'gradle')
-            }
-        }
-    }
-}
-
 def apiChangesJsonFile = project.file("src/changes/accepted-public-api-changes.json")
 def acceptedViolations = AcceptedApiChanges.parse(apiChangesJsonFile.text)
 def compatibilityBaselineVersion = moduleIdentity.releasedVersions.get().mostRecentRelease.version
