+++ build-logic/build.gradle.kts
@@ -2,7 +2,7 @@ tasks.register("check") {
     dependsOn(subprojects.map { "${it.name}:check" })
 }
 
-val clean by tasks.registering {
+tasks.register("clean") {
     val buildSrcPropertiesFile = layout.projectDirectory.file("gradle.properties")
     val rootPropertiesFile = layout.projectDirectory.file("../gradle.properties")
     doLast {
@@ -10,7 +10,7 @@ val clean by tasks.registering {
         val rootProperties = readProperties(rootPropertiesFile.asFile)
         val jvmArgs = listOf(buildSrcProperties, rootProperties).map { it.getProperty("org.gradle.jvmargs") }.toSet()
         if (jvmArgs.size > 1) {
-            throw GradleException("gradle.properties and buildSrc/gradle.properties have different org.gradle.jvmargs " +
+            throw GradleException("gradle.properties and build-logic/gradle.properties have different org.gradle.jvmargs " +
                 "which may cause two daemons to be spawned on CI and in IDEA. " +
                 "Use the same org.gradle.jvmargs for both builds.")
         }
