+++ build-logic/buildquality/src/main/kotlin/gradlebuild.incubation-report-aggregation.gradle.kts
@@ -2,7 +2,7 @@ import gradlebuild.incubation.tasks.IncubatingApiAggregateReportTask
 
 plugins {
     id("gradlebuild.lifecycle")
-    id("base")
+    id("java-base")
 }
 
 val reports by configurations.creating {
