+++ build-logic/uber-plugins/src/main/kotlin/gradlebuild.java-library.gradle.kts
@@ -17,7 +17,6 @@
 plugins {
     `java-library`
     id("gradlebuild.dependency-modules")
-    id("gradlebuild.repositories")
     id("gradlebuild.minify")
     id("gradlebuild.reproducible-archives")
     id("gradlebuild.unittest-and-compile")
