+++ subprojects/internal-android-performance-testing/build.gradle.kts
@@ -3,10 +3,6 @@ plugins {
     application
 }
 
-repositories {
-    google()
-}
-
 dependencies {
     implementation(project(":base-services"))
     implementation(project(":tooling-api"))
