+++ subprojects/internal-integ-testing/build.gradle.kts
@@ -59,7 +59,7 @@ dependencies {
     implementation(libs.ivy)
     implementation(libs.ant)
     implementation(libs.jgit) {
-        because("Some tests require a git reportitory - see AbstractIntegrationSpec.initGitDir(")
+        because("Some tests require a git repository - see AbstractIntegrationSpec.initGitDir(")
     }
     implementation(libs.jetbrainsAnnotations) {
         because("Generated language annotations for spock tests")
